Return to use the official release of markdown-ui
