+++ Core/Package.swift
@@ -37,7 +37,7 @@ let package = Package(
         .package(url: "https://github.com/raspu/Highlightr", from: "2.1.0"),
         .package(url: "https://github.com/JohnSundell/Splash", branch: "master"),
         .package(url: "https://github.com/nmdias/FeedKit", from: "9.1.2"),
-        .package(url: "https://github.com/intitni/swift-markdown-ui", branch: "main"),
+        .package(url: "https://github.com/gonzalezreal/swift-markdown-ui", from: "2.1.0"),
         .package(url: "https://github.com/sparkle-project/Sparkle", from: "2.0.0"),
     ],
     targets: [
